Moved matchers into Plugin directory and added custom Plugin annotation
