+++ src/Plugin/Matcher/TypeMatcher.php
@@ -2,11 +2,19 @@
 
 /**
  * @file
- * Contains \Drupal\rules\Matcher\TypeMatcher.
+ * Contains \Drupal\rules\Plugin\Matcher\TypeMatcher.
  */
 
-namespace Drupal\rules\Matcher;
+namespace Drupal\rules\Plugin\Matcher;
 
+/**
+ * Defines a type matcher.
+ *
+ * @RulesDataMatcher(
+ *   id = "rules_matcher_type",
+ *   label = @Translation("A type matcher.")
+ * )
+ */
 class TypeMatcher implements MatcherInterface {
 
   public function match($subject, $object) {
